refactor(file-size): split the two files closest to the 1500-line ratchet
Neither crates/stella-core/src/step.rs (1419) nor
crates/stella-pipeline/src/pipeline/tests/verification_hardening.rs (1442)
is grandfathered in scripts/file-size-baseline.txt, so each was within
sixty to eighty lines of a hard gate failure that the baseline cannot
absorb: a file that crosses 1500 without an entry must be split, and no
entry may be added for it. The next PR to add a moderate block to either
would have paid for an unrelated split to land its own change -- which is
the failure that reddened main at 9ad88824 hours ago.

Both splits are pure moves, no behaviour change:

- step.rs keeps its production half and declares its tests as the sibling
  crates/stella-core/src/step/tests.rs, the idiom already used by
  compaction, comparison, driver, ingest, loop_detect and records in this
  same crate.
- verification_hardening.rs moves two whole issue series into child
  modules, following the pattern its own flip_halt_arming and
  witness_repair_bound doc comments describe: #1291 diff coverage (with
  the scripted probe only it uses) and #1787 verdict-degradation facts.

  step.rs                 1419 -> 1099
  verification_hardening  1442 ->  973
